More improvements to the tinker tank. * Only render liquid face if liquid exists (using model for the backface) * Move GUI rendering stuff to GuiUtil (and add common smeltery/seared tank code there) * Switch tinker tank to a ticking TE for the sake of structure building. Once build ticking does nothing * Switch height reduction from checking the liquid amount to the liquid height, as we change the height but keep the amount the same (causing it to keep removing pixels from the same liquid)
