+++ resources/assets/tconstruct/blockstates/tinker_tank_controller.json
@@ -2,11 +2,6 @@
 	"forge_marker": 1,
 	"defaults": {
 		"model": "orientable",
-		"textures": {
-            "top": "tconstruct:blocks/smeltery/seared_brick",
-			"front": "tconstruct:blocks/smeltery/tinker_tank",
-            "side": "#top"
-		},
         "uvlock": true
 	},
 	"variants": {
@@ -17,12 +12,21 @@
 			"east": { "y": 90 }
 		},
 		"active": {
-			"true": { "textures": {
-				"particle": "tconstruct:blocks/smeltery/tinker_tank",
-				"front": "tconstruct:blocks/smeltery/tinker_tank_overlay"
-			} },
-			"false": {}
+			"true": { "model": "tconstruct:tinker_tank_active" },
+			"false": {
+				"textures": {
+					"top": "tconstruct:blocks/smeltery/seared_brick",
+					"front": "tconstruct:blocks/smeltery/tinker_tank",
+					"side": "#top"
+				}
+			}
 		},
-		"inventory": [{}]
+		"inventory": [{
+			"textures": {
+				"top": "tconstruct:blocks/smeltery/seared_brick",
+				"front": "tconstruct:blocks/smeltery/tinker_tank",
+				"side": "#top"
+			}
+		}]
 	}
 }
