+++ src/main/java/slimeknights/tconstruct/smeltery/client/GuiSmeltery.java
@@ -1,23 +1,15 @@
 package slimeknights.tconstruct.smeltery.client;
 
-import com.google.common.collect.Lists;
-
 import net.minecraft.client.renderer.GlStateManager;
 import net.minecraft.util.ResourceLocation;
-import net.minecraft.util.text.TextFormatting;
-import net.minecraftforge.fluids.FluidStack;
-
 import java.io.IOException;
 import java.util.List;
 
 import slimeknights.mantle.client.gui.GuiElement;
-import slimeknights.tconstruct.common.TinkerNetwork;
 import slimeknights.tconstruct.library.Util;
-import slimeknights.tconstruct.library.client.RenderUtil;
-import slimeknights.tconstruct.library.smeltery.SmelteryTank;
+import slimeknights.tconstruct.library.client.GuiUtil;
 import slimeknights.tconstruct.smeltery.client.module.GuiSmelterySideInventory;
 import slimeknights.tconstruct.smeltery.inventory.ContainerSmeltery;
-import slimeknights.tconstruct.smeltery.network.SmelteryFluidClicked;
 import slimeknights.tconstruct.smeltery.tileentity.TileSmeltery;
 import slimeknights.tconstruct.tools.inventory.ContainerSideInventory;
 
@@ -68,28 +60,9 @@ public class GuiSmeltery extends GuiHeatingStructureFuelTank {
     mouseY -= cornerY;
 
     // Liquids
-    if(8 <= mouseX && mouseX < 60 && 16 <= mouseY && mouseY < 68) {
-      FluidStack hovered = getFluidHovered(68 - mouseY - 1);
-      List<String> text = Lists.newArrayList();
-
-      if(hovered == null) {
-        int usedCap = smeltery.getTank().getFluidAmount();
-        int maxCap = smeltery.getTank().getCapacity();
-        text.add(TextFormatting.WHITE + Util.translate("gui.smeltery.capacity"));
-        RenderUtil.amountToString(maxCap, text);
-        text.add(Util.translateFormatted("gui.smeltery.capacity_available"));
-        RenderUtil.amountToString(maxCap - usedCap, text);
-      }
-      else {
-        text.add(TextFormatting.WHITE + hovered.getLocalizedName());
-        RenderUtil.liquidToString(hovered, text);
-      }
-
-      this.drawHoveringText(text, mouseX, mouseY);
-    }
-    // Fuel
-    else if(71 <= mouseX && mouseX < 83 && 16 <= mouseY && mouseY < 68) {
-      drawFuelTooltip(mouseX, mouseY);
+    List<String> tooltip = GuiUtil.drawTankTooltip(smeltery.getTank(), mouseX, mouseY, 8, 16, 60, 68);
+    if(tooltip != null) {
+      this.drawHoveringText(tooltip, mouseX, mouseY);
     }
   }
 
@@ -100,22 +73,7 @@ public class GuiSmeltery extends GuiHeatingStructureFuelTank {
     super.drawGuiContainerBackgroundLayer(partialTicks, mouseX, mouseY);
 
     // draw liquids
-    SmelteryTank liquids = smeltery.getTank();
-    if(liquids.getFluidAmount() > 0) {
-      int capacity = Math.max(liquids.getFluidAmount(), liquids.getCapacity());
-      int[] heights = calcLiquidHeights(liquids.getFluids(), capacity);
-      int x = 8 + cornerX;
-      int y = 16 + scala.h + cornerY; // y starting position
-      int w = scala.w;
-
-      for(int i = 0; i < heights.length; i++) {
-        int h = heights[i];
-        FluidStack liquid = liquids.getFluids().get(i);
-        RenderUtil.renderTiledFluid(x, y - h, w, h, this.zLevel, liquid);
-
-        y -= h;
-      }
-    }
+    GuiUtil.drawGuiTank(smeltery.getTank(), 8 + cornerX, 16 + cornerY, scala.w, scala.h, this.zLevel);
 
     // update fuel info
     fuelInfo = smeltery.getFuelDisplay();
@@ -125,43 +83,9 @@ public class GuiSmeltery extends GuiHeatingStructureFuelTank {
   @Override
   protected void mouseClicked(int mouseX, int mouseY, int mouseButton) throws IOException {
     if(mouseButton == 0) {
-      mouseX -= cornerX;
-      mouseY -= cornerY;
-      if(8 <= mouseX && mouseX < 60 && 16 <= mouseY && mouseY < 68) {
-        SmelteryTank tank = smeltery.getTank();
-        int[] heights = calcLiquidHeights(tank.getFluids(), tank.getCapacity());
-        int y = 68 - mouseY - 1;
-
-        for(int i = 0; i < heights.length; i++) {
-          if(y < heights[i]) {
-            TinkerNetwork.sendToServer(new SmelteryFluidClicked(i));
-            return;
-          }
-          y -= heights[i];
-        }
-      }
-      mouseX += cornerX;
-      mouseY += cornerY;
+      GuiUtil.handleTankClick(smeltery.getTank(), mouseX - cornerX, mouseY - cornerY, 8, 16, 60, 68);
     }
     super.mouseClicked(mouseX, mouseY, mouseButton);
-  }
-
-  protected FluidStack getFluidHovered(int y) {
-    SmelteryTank tank = smeltery.getTank();
-    int[] heights = calcLiquidHeights(tank.getFluids(), tank.getCapacity());
-
-    for(int i = 0; i < heights.length; i++) {
-      if(y < heights[i]) {
-        return tank.getFluids().get(i);
-      }
-      y -= heights[i];
-    }
-
-    return null;
-  }
-
-  protected int[] calcLiquidHeights(List<FluidStack> liquids, int capacity) {
-    return SmelteryRenderer.calcLiquidHeights(liquids, capacity, scala.h, 3);
   }
 
 }
