+++ src/main/java/slimeknights/tconstruct/smeltery/client/TinkerTankRenderer.java
@@ -18,13 +18,10 @@ import net.minecraftforge.fluids.FluidStack;
 import org.lwjgl.opengl.GL11;
 import slimeknights.tconstruct.library.client.RenderUtil;
 import slimeknights.tconstruct.library.smeltery.SmelteryTank;
-import slimeknights.tconstruct.smeltery.TinkerSmeltery;
 import slimeknights.tconstruct.smeltery.block.BlockTinkerTankController;
 import slimeknights.tconstruct.smeltery.tileentity.TileTinkerTank;
 
 public class TinkerTankRenderer extends SmelteryTankRenderer<TileTinkerTank> {
-  public static final String TEXTURE_TINKER_TANK_BACKGROUND = "tconstruct:blocks/smeltery/tinker_tank_background";
-
   protected static Minecraft mc = Minecraft.getMinecraft();
 
   @Override
@@ -45,45 +42,42 @@ public class TinkerTankRenderer extends SmelteryTankRenderer<TileTinkerTank> {
     // draw the bottom most fluid on the controller itself
     World world = tinkerTank.getWorld();
     IBlockState state = world.getBlockState(pos);
-    if(state.getBlock() == TinkerSmeltery.tinkerTankController) {
-      Tessellator tessellator = Tessellator.getInstance();
-      VertexBuffer renderer = tessellator.getBuffer();
+    // on the odd chance something outside of us changes the blockstate and uses this TE
+    if(state.getBlock() instanceof BlockTinkerTankController) {
 
-      renderer.begin(GL11.GL_QUADS, DefaultVertexFormats.BLOCK);
-      mc.renderEngine.bindTexture(TextureMap.LOCATION_BLOCKS_TEXTURE);
-
-      EnumFacing face = world.getBlockState(pos).getValue(BlockTinkerTankController.FACING);
-
-      // set up data for the sprite rendering
+      // make sure we have a fluid before continuing
       FluidStack fluidStack = tank.getFluid();
-      TextureAtlasSprite sprite;
-      int brightness, color;
       if(fluidStack != null) {
+        Tessellator tessellator = Tessellator.getInstance();
+        VertexBuffer renderer = tessellator.getBuffer();
+
+        renderer.begin(GL11.GL_QUADS, DefaultVertexFormats.BLOCK);
+        mc.renderEngine.bindTexture(TextureMap.LOCATION_BLOCKS_TEXTURE);
+
+        EnumFacing face = world.getBlockState(pos).getValue(BlockTinkerTankController.FACING);
+
+        // set up data for the sprite rendering
         Fluid fluid = fluidStack.getFluid();
-        sprite = mc.getTextureMapBlocks().getTextureExtry(fluid.getStill().toString());
-        brightness = world.getCombinedLight(pos.offset(face), fluid.getLuminosity(fluidStack));
-        color = fluid.getColor(fluidStack);
-      }
-      else {
-        // if the fluid is missing (aka empty tank) use the default texture
-        sprite = mc.getTextureMapBlocks().getTextureExtry(TEXTURE_TINKER_TANK_BACKGROUND);
-        brightness = world.getCombinedLight(pos.offset(face), 0);
-        color = 0xFFFFFFFF;
-      }
+        TextureAtlasSprite sprite = mc.getTextureMapBlocks().getTextureExtry(fluid.getStill().toString());
+        int brightness = world.getCombinedLight(pos.offset(face), fluid.getLuminosity(fluidStack));
+        int color = fluid.getColor(fluidStack);
 
-      RenderUtil.pre(x, y, z);
+        RenderUtil.pre(x, y, z);
 
-      // fluid is missing texture/default got lost
-      if(sprite == null) {
-        sprite = mc.getTextureMapBlocks().getMissingSprite();
+        // fluid is missing texture
+        if(sprite == null) {
+          sprite = mc.getTextureMapBlocks().getMissingSprite();
+        }
+
+        float d = RenderUtil.FLUID_OFFSET;
+        float d2 = 1 - (d * 2);
+        // transparent liquids should render solid textured
+        // otherwise it would look bad as our backface is so dark
+        GlStateManager.disableBlend();
+        RenderUtil.putTexturedQuad(renderer, sprite, d, d, d, d2, d2, d2, face, color, brightness, false);
+        tessellator.draw();
+        RenderUtil.post();
       }
-
-      float d = RenderUtil.FLUID_OFFSET;
-      float d2 = 1 - (d * 2);
-      GlStateManager.disableBlend(); // transparent liquids should render solid textured
-      RenderUtil.putTexturedQuad(renderer, sprite, d, d, d, d2, d2, d2, face, color, brightness, false);
-      tessellator.draw();
-      RenderUtil.post();
     }
 
     // draw the fluids inside
